refactor(condColor): extract shared helper, unify training-tab condition color with the pre-match band
CodeRabbit review comment on the training-groups Condition column: the
same `player.condition` value picked up two different color schemes
depending on the screen — the training tab flagged amber under 60,
the pre-match table used the 75/50/other band. Consistency across the
UI should trump personal thresholds; a value that looks amber in one
place should look amber everywhere.

Move `condColor` from `PreMatchLineup.tsx` (where it was awkwardly
co-located with a component) to `src/lib/playerConditionDisplay.ts`
and use it in `TrainingGroupsCard.tsx` too. Both `PreMatchSetup.tsx`
and the pre-match test file now import from the shared location.
