+++ src/components/match/PreMatchSetup.tsx
@@ -4,7 +4,8 @@ import { useTranslation } from "react-i18next";
 import { FixtureData, GameStateData } from "../../store/gameStore";
 import { getFixtureDisplayLabel } from "../../lib/helpers";
 import { MatchSnapshot, EnginePlayerData, FORMATIONS, PLAY_STYLES } from "./types";
-import PreMatchLineup, { parseFormationNeeds, POSITION_KEY_STATS, condColor, statColor, starterOvrColor, getStatVal } from "./PreMatchLineup";
+import PreMatchLineup, { parseFormationNeeds, POSITION_KEY_STATS, statColor, starterOvrColor, getStatVal } from "./PreMatchLineup";
+import { condColor } from "../../lib/playerConditionDisplay";
 import { getSetPieceStats } from "./SetPieceSelector";
 import { FormationPitch } from "./FormationPitch";
 import { makeTeamFallback } from "./helpers";
